CLN:config module cleanup
