+++ config/model.go
@@ -12,8 +12,8 @@ import (
 type ModelEmailUser struct {
 	gorm.Model
 	Email    string `gorm:"unique;not null" json:"email"`
-	Password string `gorm:"not null" json:"password"`
-	Username string `gorm:"not null" json:"username"`
+	Password string `gorm:"not null"        json:"password"`
+	Username string `gorm:"not null"        json:"username"`
 }
 
 type ModelWalletUser struct {
@@ -23,10 +23,10 @@ type ModelWalletUser struct {
 }
 
 type ModelEnvBot struct {
-	gorm.Model                
-	TriggerName        string `gorm:"not null" json:"triggerName"`
+	gorm.Model
+	TriggerName        string `gorm:"not null"        json:"triggerName"`
 	SlackBotOauthToken string `gorm:"unique;not null" json:"slackBotOauthToken"`
-	SlackChannelID     string `gorm:"not null" json:"slackChannelID"`
-	SlackUserID        string `gorm:"not null" json:"slackUserID"`
-	SlackUserName      string `gorm:"not null" json:"slackUserName"`
+	SlackChannelID     string `gorm:"not null"        json:"slackChannelID"`
+	SlackUserID        string `gorm:"not null"        json:"slackUserID"`
+	SlackUserName      string `gorm:"not null"        json:"slackUserName"`
 }
